refactor: collect human tool prompts to `prompts/`
